- feat : 음식점 별점 필드 타입 float To double

+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/dto/restaurant/RestaurantDetailDto.java
@@ -17,7 +17,7 @@ public class RestaurantDetailDto {
     private String distance;
     private String fileName;
     private int reviewCount;
-    private float starLiked;
+    private double starLiked;
     private int likedCount;
     private int viewCount;
     private boolean like;
